+++ NEWS.md
@@ -1,5 +1,7 @@
 ## ggimage 0.2.4
 
++ bug fixed of checking phylopic ID (2019-10-17, Thu)
+  - <https://groups.google.com/d/msg/bioc-ggtree/XfRV7BqNE4g/zCYH5aT5CwAJ>
 + fixed image aspect ratio (2019-10-02, Wed)
   - <https://stackoverflow.com/questions/58165226/is-it-possible-to-plot-images-in-a-ggplot2-plot-that-dont-get-distorted-when-y?noredirect=1#comment102713437_58165226>